trying to fix travis
debugging travis

continued travis debug

keep debugging

arrrrgggg travis

arrrrgggg travis

arrrrgggg travis

arrrrgggg travis

figuring out how the travis matrix works

trying to propagate env from travis matrix into scripts

trying to propagate env from travis matrix into scripts

trying to propagate env from travis matrix into scripts

trying to propagate env from travis matrix into scripts

trying to propagate env from travis matrix into scripts

trying to propagate env from travis matrix into scripts

trying to propagate env from travis matrix into scripts

trying to propagate env from travis matrix into scripts

fixed more compilation warnings in non-library sources

trying again with cleaned cache building only dependencies

after clean build of dependencies now try building again

sigh --- need static libraries -- delete cache and try again

more compiler warning suppression

more compiler warning suppression

static libraries needed for simplest linux cmake build

add configuring and enabling ccache; enable more tests

enable more tests; check that ccache is working

restoring full matrix including clang and osx

get clang working in linux

trying to get osx working again

trying to get osx working again

trying to force a build

fixing yml problem?

speed up travis tests

missing space in os:osx was breaking yml; try to use find package to fix clang osx linking problem; make run tests more verbose to stop travis from timing out

ugh; still don't understand what error is in osx exclude

missing flags on find_package

working towards working minimal set

.travis.yml

fix stupidity with path env; try to force ccache to actually cache

removing osx from travis ci matrix

make default tests tiny --- use --large argument on test to get a full test

trying to get better caching

make test sizes manageable and also controllable via env or arg

make travis build a bit faster by changing Os to O1

configure all tests with small versions controllable from env or cmdline

configure gentensor test and enable small tests

shrink tests further

+++ src/madness/mra/funcimpl.h
@@ -1362,11 +1362,11 @@ namespace madness {
 
 
                 if (NDIM == 3) {
-                    for (int i=0; i<npt; ++i) {
+                    for (size_t i=0; i<npt; ++i) {
                         c[0] = cell(0,0) + h*cell_width[0]*(l[0] + qx(i)); // x
-                        for (int j=0; j<npt; ++j) {
+                        for (size_t j=0; j<npt; ++j) {
                             c[1] = cell(1,0) + h*cell_width[1]*(l[1] + qx(j)); // y
-                            for (int k=0; k<npt; ++k) {
+                            for (size_t k=0; k<npt; ++k) {
                                 c[2] = cell(2,0) + h*cell_width[2]*(l[2] + qx(k)); // z
                                 //								fprintf(pFile,"%18.12f %18.12f %18.12f\n",c[0],c[1],c[2]);
                                 auto& success1 = std::getline(gfile,gline); MADNESS_ASSERT(success1);
